Fix meal editting problem (#58)
* FIX: Resolve ConcurrentModificationException in SendMealsToNotificationUseCase

This commit addresses a `ConcurrentModificationException` that occurred when iterating and modifying the `meals` list simultaneously in `SendMealsToNotificationUseCase`.

The fix involves using an `Iterator` to safely remove elements from the list while iterating. Additionally, the app version in `UpdatesNotesDialog.kt` has been updated to `0.7.2`.

* FIX: Resolve ConcurrentModificationException in SendMealsToNotificationUseCase

This commit addresses a `ConcurrentModificationException` that occurred when iterating and modifying the `meals` list simultaneously in `SendMealsToNotificationUseCase`.

The fix involves using an `Iterator` to safely remove elements from the list while iterating. Additionally, the app version in `UpdatesNotesDialog.kt` has been updated to `0.7.3`.