Auth feature (#65)
* REFAC: Introduce AuthService and related use cases for Firebase authentication

This commit introduces an `AuthService` to handle Firebase authentication, specifically anonymous sign-in and fetching the user ID. It also adds corresponding use cases (`SignInAnonymouslyUseCase`, `GetUserIdUseCase`) and refactors the `VersionManageRepository` by moving it to a new `repositories` package.

Key changes:

- Created `AuthService.kt` to encapsulate Firebase authentication logic.
- Implemented `signInAnonymously()` and `getUid()` methods in `AuthService`.
- Added `SignInAnonymouslyUseCase.kt` to provide a use case for anonymous sign-in.
- Added `GetUserIdUseCase.kt` to provide a use case for retrieving the current user's ID.
- Created `AuthRepository.kt` to act as an intermediary between the use cases and `AuthService`.
- Moved `VersionManageRepository.kt` to `app/src/main/java/com/maverickapps/nutripet/core/data/repositories/`.
- Updated import statements in `FetchCurrentVerUseCase.kt` and `CheckCurrentVerIsLatestUseCase.kt` to reflect the new location of `VersionManageRepository`.
- Removed `TryAuthUseCase.kt` as its functionality is now covered by the new authentication components.
- Updated `FirebaseModule.kt` to provide an instance of `AuthService`.
- Removed `TryAuthUseCase` dependency from `DashboardViewModel.kt`.

* REFAC: Make AuthRepository.signInAnonymously suspending

This commit modifies the `AuthRepository.signInAnonymously` function to be a suspending function. It also updates the function to return "e" as a default value if the anonymous sign-in fails or returns null.

Additionally, the unused `TryAuthUseCase` dependency was removed from `DashboardViewModel`.

* CHORE: Add suspend modifier to SignInAnonymouslyUseCase

This commit adds the `suspend` modifier to the `invoke` operator function in `SignInAnonymouslyUseCase`. This change aligns the function with asynchronous operations performed by the `AuthRepository`.